+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @luwy-dyro/ui
 
+## 1.1.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @luwy-dyro/tokens@0.8.0
+
+## 1.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @luwy-dyro/tokens@0.7.0
+
 ## 1.1.7
 
 ### Patch Changes